+++ src/piper/__main__.py
@@ -10,11 +10,12 @@ import wave
 from collections.abc import Iterable
 from pathlib import Path
 
+from pathvalidate import sanitize_filename
+
 from . import PiperVoice, SynthesisConfig
 from .audio_playback import AudioPlayer
 
 _FILE = Path(__file__)
-_DIR = _FILE.parent
 _LOGGER = logging.getLogger(_FILE.stem)
 
 
@@ -41,6 +42,12 @@ def main() -> None:
         "--output-dir",
         "--output_dir",
         help="Path to output directory (default: cwd)",
+    )
+    parser.add_argument(
+        "--output-dir-naming",
+        choices=("timestamp", "text"),
+        default="timestamp",
+        help="Naming scheme for WAV files in output directory (default: timestamp)",
     )
     parser.add_argument(
         "--output-raw",
@@ -187,10 +194,28 @@ def main() -> None:
         output_dir.mkdir(parents=True, exist_ok=True)
 
         for line in lines():
-            wav_path = output_dir / f"{time.monotonic_ns()}.wav"
+            if args.output_dir_naming == "text":
+                # WAV file is named after input text (may overwrite).
+                text = sanitize_filename(line.strip())
+                wav_path = output_dir / f"{text}.wav"
+            else:
+                # WAV file is named with a unique timestamp
+                wav_path = output_dir / f"{time.monotonic_ns()}.wav"
+
             wav_file = wave.open(str(wav_path), "wb")
             with wav_file:
-                lines_to_wav()
+                wav_params_set = False
+                for i, audio_chunk in enumerate(voice.synthesize(line, syn_config)):
+                    if not wav_params_set:
+                        wav_file.setframerate(audio_chunk.sample_rate)
+                        wav_file.setsampwidth(audio_chunk.sample_width)
+                        wav_file.setnchannels(audio_chunk.sample_channels)
+                        wav_params_set = True
+
+                    if i > 0:
+                        wav_file.writeframes(silence_int16_bytes)
+
+                    wav_file.writeframes(audio_chunk.audio_int16_bytes)
 
             _LOGGER.info("Wrote %s", wav_path)
     else: